requirements.txt
# Jung Claude System v1.0 - Dependências

# Core AI
anthropic>=0.7.0
openai>=1.3.0
langchain>=0.1.0
langchain-anthropic>=0.1.0
langchain-openai>=0.1.0
langchain-chroma>=0.1.0

# Vector Database
chromadb>=0.4.0

# Web Interface  
streamlit>=1.28.0
streamlit-chat>=0.1.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
asyncio-throttle>=1.0.0
aiofiles>=23.0.0
tenacity>=8.0.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0